requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-pubsub==2.23.0
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.8
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.32.3
rsa==4.9
urllib3==2.2.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-pubsub==2.23.0
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.8
proto-plus==1.24.0
protobuf==5.28.0
py4j==0.10.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyspark==3.5.2
requests==2.32.3
rsa==4.9
urllib3==2.2.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.23.0
google-cloud-translate==3.16.0
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.8
proto-plus==1.24.0
protobuf==5.28.0
py4j==0.10.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyspark==3.5.2
requests==2.32.3
rsa==4.9
urllib3==2.2.2
